requirements.txt
asgiref==3.11.0
Django==4.2.27
sqlparse==0.5.5
typing_extensions==4.15.0
faker-37.12.0